feat: 설정 화면 layout 구현
