--- app/src/main/java/woowacourse/movie/SettingFragment.kt
@@ -1,17 +1,0 @@
-package woowacourse.movie
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-
-class SettingFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        return inflater.inflate(R.layout.fragment_setting, container, false)
-    }
-}
