+++ app/src/main/java/woowacourse/movie/setting/SettingFragment.kt
@@ -1,0 +1,24 @@
+package woowacourse.movie.setting
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.fragment.app.Fragment
+import woowacourse.movie.R
+import woowacourse.movie.databinding.FragmentSettingBinding
+
+class SettingFragment : Fragment() {
+    private var _binding: FragmentSettingBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_setting, container, false)
+        return binding.root
+    }
+}
